Subdiv and volumes

+++ Spurious/SpuriousApi/Models/AlcoholVolumes.cs
@@ -6,9 +6,9 @@ namespace SpuriousApi.Models
 {
     public class AlcoholVolumes
     {
-        public int Beer { get; set; }
-        public int Wine { get; set; }
-        public int Spirits { get; set; }
-        public int Total { get { return this.Beer + this.Wine + this.Spirits; } }
+        public long Beer { get; set; }
+        public long Wine { get; set; }
+        public long Spirits { get; set; }
+        public long Total { get { return this.Beer + this.Wine + this.Spirits; } }
     }
 }
